requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
qrcode==7.4.2
pillow==10.1.0
pydantic==2.4.2
python-dotenv==1.0.0
supabase==1.0.4